move flightdeck dir and files - flightdeck files now live at the root directory
