+++ .flightdeck.js
@@ -1,10 +1,10 @@
 // Eleventy Configuration
-const addWorkflow = require("./src/_flightdeck/workflow");
-const addFilters = require("./src/_flightdeck/filters");
-const addShortcodes = require("./src/_flightdeck/shortcodes");
-const addComponents = require("./src/_flightdeck/components");
-const addTransforms = require("./src/_flightdeck/transforms");
-const addPlugins = require("./src/_flightdeck/plugins");
+const addWorkflow = require("./.flightdeck/workflow");
+const addFilters = require("./.flightdeck/filters");
+const addShortcodes = require("./.flightdeck/shortcodes");
+const addComponents = require("./.flightdeck/components");
+const addTransforms = require("./.flightdeck/transforms");
+const addPlugins = require("./.flightdeck/plugins");
 const isProd = process.env.ENVIRONMENT === "prod";
 
 module.exports = (config) => {
