+++ .flightdeck/shortcodes/copyright.js
