+++ .flightdeck/transforms/buildImages.js
