+++ .vscode/settings.json
@@ -29,5 +29,8 @@
       "scope": "workspace"
     }
   },
-  "cSpell.words": ["Laserwave"]
+  "cSpell.words": [
+    "flightdeck",
+    "Laserwave"
+  ]
 }
